[refactorrefpattern] substituted ref patterns: now its debug time
