+++ Refine/TPZRefPatternTools.cpp
@@ -10,7 +10,8 @@
 
 #include "TPZRefPatternTools.h"
 #include "TPZVTKGeoMesh.h"
-#include "TPZRefPattern3.h"//@TODOFran: remove this
+#include "TPZRefPattern.h"
+#include "TPZRefPatternDataBase.h"
 
 using namespace std;
 
@@ -849,52 +850,10 @@ std::string TPZRefPatternTools::BuildRefPatternModelName(TPZRefPattern &refp)
 	
 	if(refpTypeName.length() == 0)
 	{
-		refpTypeName = nonInitializedName;
+		refpTypeName = TPZRefPattern::fNonInitializedName;
 	}
 	
 	return refpTypeName;
-}
-
-std::string TPZRefPatternTools::BuildRefPatternModelName(TPZRefPattern3 &refp)
-{
-    std::string refpTypeName;
-    std::stringstream rpName;
-    TPZRefPattern3 *prefp = &refp;
-
-    if(prefp == NULL)
-    {
-        std::cout << "Null refpattern parameter on " << __PRETTY_FUNCTION__ << std::endl;
-        return refpTypeName;
-    }
-
-    TPZGeoEl *gel = refp.Element(0);
-    int ncorners = gel->NCornerNodes();
-    int nsides = gel->NSides();
-
-    std::string perfix = gel->TypeName();
-    for(int i = 0; i < 3; i++)
-    {
-        rpName <<  perfix[i];
-    }
-    for(int s = 0; s < nsides; s++)
-    {
-        if(s < ncorners)
-        {
-            rpName << "0";
-        }
-        else
-        {
-            rpName << refp.NSideNodes(s);
-        }
-    }
-    rpName >> refpTypeName;
-
-    if(refpTypeName.length() == 0)
-    {
-        refpTypeName = nonInitializedName;
-    }
-
-    return refpTypeName;
 }
 
 std::string TPZRefPatternTools::BuildRefPatternModelName(TPZAutoPointer<TPZRefPattern> refp)
@@ -932,7 +891,7 @@ std::string TPZRefPatternTools::BuildRefPatternModelName(TPZAutoPointer<TPZRefPa
 	
 	if(refpTypeName.length() == 0)
 	{
-		refpTypeName = nonInitializedName;
+		refpTypeName = TPZRefPattern::fNonInitializedName;
 	}
 	
 	return refpTypeName;
@@ -971,7 +930,7 @@ std::string TPZRefPatternTools::BuildRefPatternModelName(TPZGeoEl *gel)
 	
 	if(refpTypeName.length() == 0)
 	{
-		refpTypeName = nonInitializedName;
+		refpTypeName = TPZRefPattern::fNonInitializedName;
 	}
 	
 	return refpTypeName;
