+++ Refine/TPZRefPatternTools.h
@@ -18,10 +18,9 @@
 #include <set>
 
 #include "pzgeoel.h"
-#include "TPZRefPattern.h"
 #include "tpzintpoints.h"
 
-class TPZRefPattern3;
+class TPZRefPattern;
 /**
  * @ingroup refine
  * @brief Defines tools of pattern. \ref refine "Refine"
@@ -145,7 +144,6 @@ public:
 	 * followed by the quantity of midnodes for each side of element.
 	 */
 	static std::string BuildRefPatternModelName(TPZRefPattern &refp);
-    static std::string BuildRefPatternModelName(TPZRefPattern3 &refp);
 	static std::string BuildRefPatternModelName(TPZAutoPointer<TPZRefPattern> refp);
 	static std::string BuildRefPatternModelName(TPZGeoEl *gel);
